Adds unit tests to DecorationView UIView
Adds a large number of unit tests to DecorationView since it  provides
the main view for a number of the dynamic VCs

Refactors DecorationView to accommodate more unit tests

Protects those attr_accessors that do not provide API access to the
class instance. These attr_accessorscould be replaced by ivars.

Adds a cleanup method for setting spec windoww and controller to nil.
This does not appear to be working as expected - further investigation
required.